requirements2.txt
selenium
undetected-chromedriver
python-dotenv
Faker 